visual-tree-search-state-reset/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
supabase>=1.0.3
httpx>=0.24.1
sqlalchemy>=2.0.27
pymysql>=1.1.0
docker